Move Threhold and serdes functions to params/
