+++ rtengine/CMakeLists.txt
@@ -1,3 +1,8 @@
+set(PROCPARAM_SRCS
+    procparams.cc
+    params/serdes.cc
+)
+
 set(RTENGINESOURCEFILES
     alpha.cc
     ahd_demosaic_RT.cc
@@ -91,7 +96,6 @@ set(RTENGINESOURCEFILES
     pixelshift.cc
     previewimage.cc
     processingjob.cc
-    procparams.cc
     profilestore.cc
     rawflatfield.cc
     rawimage.cc
@@ -110,6 +114,7 @@ set(RTENGINESOURCEFILES
     utils.cc
     vng4_demosaic_RT.cc
     xtrans_demosaic.cc
+    ${PROCPARAM_SRCS}
 )
 
 if(LENSFUN_HAS_LOAD_DIRECTORY)
@@ -122,18 +127,22 @@ endif()
 if(CMAKE_CXX_COMPILER_ID STREQUAL "GNU" AND CMAKE_CXX_COMPILER_VERSION VERSION_GREATER_EQUAL "12.0")
     # procparams.cc takes a long time to compile with optimizations starting
     # with GCC 12.1 due to PTA (see issue #6548)
-    get_source_file_property(PROCPARAMS_COMPILE_OPTIONS procparams.cc COMPILE_OPTIONS)
-    if(PROCPARAMS_COMPILE_OPTIONS STREQUAL "NOTFOUND")
-        set(PROCPARAMS_COMPILE_OPTIONS "")
-    else()
-        set(PROCPARAMS_COMPILE_OPTIONS "${PROCPARAMS_COMPILE_OPTIONS};")
-    endif()
-    set(PROCPARAMS_COMPILE_OPTIONS "${PROCPARAMS_COMPILE_OPTIONS}-fno-tree-pta")
+    foreach(SRC_FILE PROCPARAM_SRCS)
+        message("Processing file: ${SRC_FILE}")
 
-    set_source_files_properties(
-        procparams.cc PROPERTIES
-        COMPILE_OPTIONS ${PROCPARAMS_COMPILE_OPTIONS}
-    )
+        get_source_file_property(PROCPARAMS_COMPILE_OPTIONS ${SRC_FILE} COMPILE_OPTIONS)
+        if(PROCPARAMS_COMPILE_OPTIONS STREQUAL "NOTFOUND")
+            set(PROCPARAMS_COMPILE_OPTIONS "")
+        else()
+            set(PROCPARAMS_COMPILE_OPTIONS "${PROCPARAMS_COMPILE_OPTIONS};")
+        endif()
+        set(PROCPARAMS_COMPILE_OPTIONS "${PROCPARAMS_COMPILE_OPTIONS}-fno-tree-pta")
+
+        set_source_files_properties(
+            ${SRC_FILE} PROPERTIES
+            COMPILE_OPTIONS ${PROCPARAMS_COMPILE_OPTIONS}
+        )
+    endforeach()
 endif()
 
 if(NOT WITH_SYSTEM_KLT)
